@@ -5,10 +5,13 @@
   "dependencies": {
     "@apollo/client": "^3.3.20",
     "@apollo/react-hooks": "^3.1.3",
+    "@emotion/react": "^11.11.0",
+    "@emotion/styled": "^11.11.0",
     "@googlemaps/markerclusterer": "^1.0.16",
-    "@material-ui/core": "^4.8.0",
-    "@material-ui/icons": "^4.5.1",
-    "@material-ui/lab": "^4.0.0-alpha.36",
+    "@mui/icons-material": "^5.11.16",
+    "@mui/lab": "^5.0.0-alpha.132",
+    "@mui/material": "^5.13.3",
+    "@mui/styles": "^5.13.2",
     "@react-google-maps/api": "^2.7.0",
     "@sentry/integrations": "^6.13.3",
     "@sentry/react": "^6.13.3",
@@ -22,9 +25,9 @@
     "graphql": "^14.6.0",
     "material-ui-popup-state": "^1.5.3",
     "moment": "^2.29.2",
-    "notistack": "^0.9.7",
+    "notistack": "^3.0.1",
     "prop-types": "^15.7.2",
-    "react": "^16.10.2",
+    "react": "^17.0.0",
     "react-dom": "^16.10.2",
     "react-google-autocomplete": "^2.5.0",
     "react-mobile-picker-scroll": "^0.2.14",
