@@ -1,6 +1,7 @@
 import React from 'react'
-import { Fab, makeStyles } from '@material-ui/core'
-import NearMeIcon from '@material-ui/icons/NearMe'
+import { Fab } from '@mui/material'
+import { makeStyles } from '@mui/styles'
+import NearMeIcon from '@mui/icons-material/NearMe'
 
 const useStyles = makeStyles({
   fab: {
