@@ -1,9 +1,9 @@
 import React from 'react'
 
-import Paper from '@material-ui/core/Paper'
-import Box from '@material-ui/core/Box'
-import { makeStyles } from '@material-ui/core/styles'
-import Typography from '@material-ui/core/Typography'
+import Paper from '@mui/material/Paper'
+import Box from '@mui/material/Box'
+import { makeStyles } from '@mui/styles'
+import Typography from '@mui/material/Typography'
 
 import { useMissionValue } from '../../../../../utils/contexts/MissionContext'
 import Flag from '../../Flag'
