@@ -1,19 +1,19 @@
 import React, { useState, useRef, useMemo, useEffect } from 'react'
 
-import Typography from '@material-ui/core/Typography'
-import NativeSelect from '@material-ui/core/NativeSelect'
-import Grid from '@material-ui/core/Grid'
-import TextField from '@material-ui/core/TextField'
-import { Input } from '@material-ui/core'
-import Box from '@material-ui/core/Box'
-import IconButton from '@material-ui/core/IconButton'
-import LocationOnIcon from '@material-ui/icons/LocationOn'
-import CreateIcon from '@material-ui/icons/Create'
-import StorefrontIcon from '@material-ui/icons/Storefront'
-import BusinessIcon from '@material-ui/icons/Business'
-import ApartmentIcon from '@material-ui/icons/Apartment'
-import DescriptionIcon from '@material-ui/icons/Description'
-import AddAPhotoIcon from '@material-ui/icons/AddAPhoto'
+import Typography from '@mui/material/Typography'
+import NativeSelect from '@mui/material/NativeSelect'
+import Grid from '@mui/material/Grid'
+import TextField from '@mui/material/TextField'
+import { Input } from '@mui/material'
+import Box from '@mui/material/Box'
+import IconButton from '@mui/material/IconButton'
+import LocationOnIcon from '@mui/icons-material/LocationOn'
+import CreateIcon from '@mui/icons-material/Create'
+import StorefrontIcon from '@mui/icons-material/Storefront'
+import BusinessIcon from '@mui/icons-material/Business'
+import ApartmentIcon from '@mui/icons-material/Apartment'
+import DescriptionIcon from '@mui/icons-material/Description'
+import AddAPhotoIcon from '@mui/icons-material/AddAPhoto'
 import tagData from '../../../../../../constants/tagData'
 import {
   useMissionValue,
@@ -77,7 +77,13 @@ function MissionStep2() {
     <>
       <Grid container spacing={3}>
         {/* * ==================== 1.經緯度標註 ==================== */}
-        <Grid container item xs={12} justify='space-between' direction='row'>
+        <Grid
+          container
+          item
+          xs={12}
+          justifyContent='space-between'
+          direction='row'
+        >
           <Box
             display='flex'
             flexDirection='row'
@@ -97,7 +103,13 @@ function MissionStep2() {
         </Grid>
 
         {/* * ==================== 2.地點資訊 ==================== */}
-        <Grid container item xs={12} justify='space-between' direction='row'>
+        <Grid
+          container
+          item
+          xs={12}
+          justifyContent='space-between'
+          direction='row'
+        >
           <Box
             display='flex'
             flexDirection='row'
@@ -133,7 +145,13 @@ function MissionStep2() {
           </Box>
         </Grid>
 
-        <Grid container item xs={12} justify='space-between' direction='row'>
+        <Grid
+          container
+          item
+          xs={12}
+          justifyContent='space-between'
+          direction='row'
+        >
           <Box
             display='flex'
             flexDirection='row'
