@@ -1,7 +1,7 @@
 import React from 'react'
-import { makeStyles } from '@material-ui/core/styles'
-import AddSharpIcon from '@material-ui/icons/AddSharp'
-import { Fab } from '@material-ui/core'
+import { makeStyles } from '@mui/styles'
+import AddSharpIcon from '@mui/icons-material/AddSharp'
+import { Fab } from '@mui/material'
 import { useMissionValue } from '../../../../../utils/contexts/MissionContext'
 import { useUserValue } from '../../../../../utils/contexts/UserContext'
 
