@@ -1,17 +1,17 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Popover from '@material-ui/core/Popover'
-import List from '@material-ui/core/List'
-import ListItem from '@material-ui/core/ListItem'
-import ListItemIcon from '@material-ui/core/ListItemIcon'
-import ListItemText from '@material-ui/core/ListItemText'
-import PersonIcon from '@material-ui/icons/Person'
-import DescriptionIcon from '@material-ui/icons/Description'
-import HelpIcon from '@material-ui/icons/Help'
-import ExitToAppIcon from '@material-ui/icons/ExitToApp'
-import RateReviewIcon from '@material-ui/icons/RateReview'
-import { makeStyles } from '@material-ui/core/styles'
+import Popover from '@mui/material/Popover'
+import List from '@mui/material/List'
+import ListItem from '@mui/material/ListItem'
+import ListItemIcon from '@mui/material/ListItemIcon'
+import ListItemText from '@mui/material/ListItemText'
+import PersonIcon from '@mui/icons-material/Person'
+import DescriptionIcon from '@mui/icons-material/Description'
+import HelpIcon from '@mui/icons-material/Help'
+import ExitToAppIcon from '@mui/icons-material/ExitToApp'
+import RateReviewIcon from '@mui/icons-material/RateReview'
+import { makeStyles } from '@mui/styles'
 
 import { useUserValue } from '../../../../utils/contexts/UserContext'
 
