@@ -1,4 +1,4 @@
-import { createMuiTheme } from '@material-ui/core/styles'
+import { createTheme } from '@mui/material/styles'
 
 const ThemeSettings = {
   palette: {
@@ -29,4 +29,4 @@ const ThemeSettings = {
   }
 }
 
-export const theme = createMuiTheme(ThemeSettings)
+export const theme = createTheme(ThemeSettings)
