@@ -20,7 +20,22 @@ import { getItem, setItem, TOKEN_EXPIRE_INFO } from '../functions/localStorage'
 const REFRESH_TOKEN_TIMEOUT = 15 * 60 * 1000 // 15 minutes
 
 // Firebase Google authentication settings
-const firebaseApp = firebase.initializeApp(firebaseConfig)
+// const firebaseApp = firebase.initializeApp(firebaseConfig)
+const firebaseApp = VITE_FIREBASE_LOCAL_SERVER
+  ? firebase.initializeApp({
+      apiKey: 'API_KEY',
+      authDomain: 'PROJECT_ID.firebaseapp.com',
+      // The value of `databaseURL` depends on the location of the database
+      databaseURL: 'https://DATABASE_NAME.firebaseio.com',
+      projectId: 'PROJECT_ID',
+      storageBucket: 'PROJECT_ID.appspot.com',
+      messagingSenderId: 'SENDER_ID',
+      appId: 'APP_ID',
+      // For Firebase JavaScript SDK v7.20.0 and later, `measurementId` is an optional field
+      measurementId: 'G-MEASUREMENT_ID'
+    })
+  : firebase.initializeApp(firebaseConfig)
+
 const firebaseAppAuth = firebaseApp.auth()
 const providers = {
   googleProvider: new firebase.auth.GoogleAuthProvider(),
