@@ -22,12 +22,10 @@ const userDetailInitial = {
 }
 
 function useUserDetail() {
-  const [
-    getUserDetail,
-    { data: { getUserData = {} } = {}, loading }
-  ] = useLazyQuery(GET_USER_DETAIL_QUERY, {
-    fetchPolicy: 'no-cache'
-  })
+  const [getUserDetail, { data: { getUserData = {} } = {}, loading }] =
+    useLazyQuery(GET_USER_DETAIL_QUERY, {
+      fetchPolicy: 'no-cache'
+    })
   const userDetail = useMemo(
     () => ({
       ...userDetailInitial,
